Revert "♻️ refactor(webview): move webview HTML generation to WebviewHTMLMana…"
This reverts commit e70954f3de2fc414d678a1c0c0da71a2a023d41b.
